Fixed bug in vdz water table calc.
In my previous runs, I was not seeing any variability in WTD at depth
underneath the upslope area. This was partially  because of the looping
index in kkloop. It was impossible to return a value of 2 as the
saturated layer. The arrmed array was changed to correct this.

In addition, to help comprehensibility of the fortran code, I changed
the last loop, which maps the index of the topmost saturated layer
(which has all completely saturated layers below it), to take
an array mapping layer number to depth as an input. Before, the input
was the layer depth array. It was confusing to take backwards cumulative
sums from that input.